requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
attrs==23.2.0
autopep8==2.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.8
distro==1.9.0
docstring-to-markdown==0.15
docx2pdf==0.1.8
dropbox==11.36.2
flake8==7.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
isort==5.13.2
jedi==0.19.1
lxml==5.1.0
mccabe==0.7.0
oauth2client==4.1.3
outcome==1.3.0.post0
packaging==24.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
ply==3.11
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydocstyle==6.3.0
pyflakes==3.2.0
pylint==3.1.0
pyparsing==3.1.2
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
PySocks==1.7.1
python-docx==1.1.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.1
pytoolconfig==1.3.1
pywin32==306
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
rope==1.12.0
rsa==4.9
setuptools==69.2.0
shiboken6==6.6.2
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
stone==3.3.1
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.1
whatthepatch==1.0.5
wsproto==1.2.0
yapf==0.40.2
zipp==3.18.1
